+++ src/components/QtubeAdmin.vue
@@ -1,10 +1,11 @@
 <script setup>
 import { mapStores } from 'pinia'
 import {useQordialAuthStore, JsonModal, NameInput, PrettyBytes, PrettyIdentifier, PrettyTime, ResourceDownloader} from 'qordial'
-import VCodeBlock from '@wdns/vue-code-block'
 import QtubeVideoThumbnail from './QtubeVideoThumbnail.vue'
+import QtubeVideoDialog from './QtubeVideoDialog.vue'
 import QtubeImagePicker from './QtubeImagePicker.vue'
 import FrameExtractor from './FrameExtractor.vue'
+import QtubeVideoList from './QtubeVideoList.vue'
 </script>
 
 <script>
@@ -12,16 +13,19 @@ export default {
 
     data() {
         return {
+            // TODO: this should come from somewhere else
             qtubeIdentifierPrefix: 'qtube_vid_',
+
+            showingNameOption: '',
+            showingName: null,
 
             qtubeVideosLimit: 20,
             qtubeVideosData: null,
+            qtubeVideosDataFinal: null,
             qtubeVideosLoading: false,
 
             qtubeVideoShowDocument: false,
-            qtubeVideoDocument: null,
-            qtubeVideoFileSize: null,
-            qtubeVideoShowJSON: null,
+            qtubeViewAfterCancelEdit: false,
 
             qtubeVideoEditing: false,
             qtubeVideoEditingResourceName: null,
@@ -45,6 +49,12 @@ export default {
 
             videoResourceShowMetadata: false,
             videoResourceMetadata: null,
+
+            qtubeVideoCache: null,
+            qtubeVideoCacheBuilding: false,
+            qtubeVideoCacheBuildingStep: null,
+            qtubeVideoCacheBuildingTotal: null,
+            videoResourcesQtubeFilter: null,
         }
     },
 
@@ -56,7 +66,30 @@ export default {
         },
     },
 
+    mounted() {
+        if (!this.showingName && this.qordialAuthStore.username) {
+            this.showingName = this.qordialAuthStore.username
+            this.showingNameOption = this.qordialAuthStore.username
+        }
+    },
+
     watch: {
+
+        'qordialAuthStore.username' (to, from) {
+            if (to) {
+                this.showingName = to
+                this.showingNameOption = to
+            }
+        },
+
+        showingName (to, from) {
+            this.qtubeVideosData = null
+            this.qtubeVideosDataFinal = null
+            this.qtubeVideoCache = null
+            this.videoResourcesData = null
+            this.videoResourcesQtubeFilter = null
+        },
+
         coverImageSource (to, from) {
             if (to == 'extract1') {
                 this.qtubeVideoEditingObject.videoImage = this.qtubeVideoEditingExtract1
@@ -73,6 +106,57 @@ export default {
     },
 
     methods: {
+
+        async rebuildQtubeCache() {
+            this.qtubeVideoCacheBuilding = true
+
+            const qtubeResources = await qortalRequest({
+                action: 'SEARCH_QDN_RESOURCES',
+                mode: 'ALL',
+                service: 'DOCUMENT',
+                name: this.showingName,
+                exactMatchNames: true,
+                identifier: this.qtubeIdentifierPrefix,
+                limit: 1000,     // TODO
+            })
+
+            // TODO: would be nice to show progress bar for this operation
+            this.qtubeVideoCache = {}
+            if (qtubeResources.length) {
+                this.qtubeVideoCacheBuildingTotal = qtubeResources.length
+                this.qtubeVideoCacheBuildingStep = 0
+                for (let info of qtubeResources) {
+                    const json = await this.$qordial.fetchResourceObject(info)
+                    const video = json.videoReference
+                    if (video && video.name == info.name) {
+                        this.qtubeVideoCache[video.identifier] = json
+                    }
+                    this.qtubeVideoCacheBuildingStep++
+                }
+            }
+
+            this.qtubeVideoCacheBuilding = false
+        },
+
+        async showingNameOptionUpdated(value) {
+            if (value === null) {
+                if (await this.$qordial.authenticate()) {
+                    value = this.qordialAuthStore.username
+                    this.showingName = value
+                    this.showingNameOption = value
+                    return
+                } else {
+                    this.showingNameOption = '' // other
+                    return
+                }
+            }
+            this.showingName = value
+            if (!value) {
+                this.$nextTick(() => {
+                    this.$refs.showingName.$el.querySelector('input').focus()
+                })
+            }
+        },
 
         async magicExtract() {
 
@@ -114,12 +198,14 @@ export default {
 
         async qtubeVideosRefresh() {
             this.qtubeVideosLoading = true
-            const url = `/arbitrary/resources/search?mode=ALL&service=DOCUMENT&name=${this.qordialAuthStore.username}&identifier=${this.qtubeIdentifierPrefix}&limit=${this.qtubeVideosLimit}&includemetadata=false&reverse=true&excludeblocked=true&exactmatchnames=true`
+            const url = `/arbitrary/resources/search?mode=ALL&service=DOCUMENT&name=${this.showingName}&identifier=${this.qtubeIdentifierPrefix}&limit=${this.qtubeVideosLimit}&includemetadata=false&reverse=true&excludeblocked=true&exactmatchnames=true`
             const response = await fetch(url)
             this.qtubeVideosData = await response.json()
             this.qtubeVideosLoading = false
         },
 
+        // (?)
+        // TODO: this is duplicated in QtubeVideoDialog
         async fetchVideoSize(resource) {
             if (resource) {
                 const response = await qortalRequest({
@@ -134,21 +220,21 @@ export default {
             }
         },
 
-        async showQtubeVideoDocument(resource) {
-            this.qtubeVideoDocumentResource = resource
-            this.qtubeVideoDocument = await this.$qordial.fetchResourceObject(resource)
-            this.qtubeVideoShowJSON = false
-            this.qtubeVideoFileSize = await this.fetchVideoSize(
-                this.qtubeVideoDocument?.videoReference)
-            this.qtubeVideoShowDocument = true
+        async showQtubeVideoDocument(info) {
+            let allowEdit = false
+            if (this.qordialAuthStore.username && info.name == this.qordialAuthStore.username) {
+                allowEdit = true
+            }
+            await this.$refs.qtubeViewDialog.showResourceInfo(info, allowEdit)
         },
 
-        async qtubeVideoInitEdit(resource, videoResource) {
+        async qtubeVideoInitEdit(resource, videoResource, viewAfterCancel = false) {
 
             // just in case view dialog is open, close it
             // TODO: should maybe check first, and re-show view dialog
             // if user cancels edit
             this.qtubeVideoShowDocument = false
+            this.qtubeViewAfterCancelEdit = viewAfterCancel
 
             // establish what we're editing
             if (resource) {
@@ -216,6 +302,15 @@ export default {
             this.$nextTick(() => {
                 this.$refs.qtubeEditTop.scrollIntoView()
             })
+        },
+
+        qtubeVideoCancelEdit() {
+            // nb. hide edit modal but then show view modal
+            this.qtubeVideoEditing = false
+
+            if (this.qtubeViewAfterCancelEdit) {
+                this.qtubeVideoShowDocument = true
+            }
         },
 
         async qtubeVideoPublish() {
@@ -289,28 +384,33 @@ export default {
             this.qtubeVideoEditing = false
         },
 
-        async downloadResource(resource) {
-            return await this.$refs.downloader.downloadResource(resource)
-        },
-
-        async viewInQtube(resource) {
-            await qortalRequest({
-                action: 'OPEN_NEW_TAB',
-                qortalLink: `qortal://APP/Q-Tube/video/${resource.name}/${resource.identifier}`,
-            })
-        },
-
         async videoResourcesRefresh() {
             this.videoResourcesLoading = true
 
-            const response = await qortalRequest({
+            this.videoResourcesData = await qortalRequest({
                 action: 'LIST_QDN_RESOURCES',
                 service: 'VIDEO',
-                name: this.qordialAuthStore.username,
+                name: this.showingName,
                 limit: this.videoResourcesLimit,
+                includeMetadata: true,
             })
 
-            this.videoResourcesData = response
+            if (this.qtubeVideoCache) {
+                for (let info of this.videoResourcesData) {
+                    info.qtubeDocument = this.qtubeVideoCache[info.identifier]
+                }
+
+                if (this.videoResourcesQtubeFilter == 'qtube-only') {
+                    this.videoResourcesData = this.videoResourcesData.filter(info => {
+                        return !!info.qtubeDocument
+                    })
+                } else if (this.videoResourcesQtubeFilter == 'qtube-missing') {
+                    this.videoResourcesData = this.videoResourcesData.filter(info => {
+                        return !info.qtubeDocument
+                    })
+                }
+            }
+
             this.videoResourcesLoading = false
         },
 
@@ -330,11 +430,22 @@ export default {
 
     <o-field grouped>
       <o-field label="Name">
-        <NameInput />
+        <div style="display: flex; gap: 1rem;">
+          <o-select v-model="showingNameOption"
+                    @update:model-value="showingNameOptionUpdated">
+            <option :value="qordialAuthStore.username">
+              {{ qordialAuthStore.username || "(please authenticate)" }}
+            </option>
+            <option :value="''">other</option>
+          </o-select>
+          <o-input v-show="!showingNameOption"
+                   v-model="showingName"
+                   ref="showingName" />
+        </div>
       </o-field>
     </o-field>
 
-    <div v-if="qordialAuthStore.username">
+    <div>
 
       <h3 class="is-size-3 has-text-primary"
           style="margin-top: 3rem;">
@@ -347,160 +458,31 @@ export default {
         </o-field>
       </o-field>
 
-      <div style="display: flex; align-items: center; gap: 2rem;">
-        <o-button @click="qtubeVideosRefresh()"
-                  :disabled="qtubeVideosLoading">
-          {{ qtubeVideosLoading ? "Working, please wait..." : (qtubeVideosData ? "Refresh" : "Load") }}
-        </o-button>
-        <span v-if="qtubeVideosData">
-          found {{ qtubeVideosData.length }} result{{ qtubeVideosData.length == 1 ? '' : 's' }}
-        </span>
-      </div>
-
-      <o-table v-if="qtubeVideosData"
-               :data="qtubeVideosData || []"
-               hoverable
-               :loading="qtubeVideosLoading">
-        <o-table-column label="Service"
-                        v-slot="{ row }">
-          {{ row.service }}
-        </o-table-column>
-        <o-table-column label="Name"
-                        v-slot="{ row }">
-          {{ row.name }}
-        </o-table-column>
-        <o-table-column label="Identifier"
-                        v-slot="{ row }">
-          <PrettyIdentifier :value="row.identifier" />
-        </o-table-column>
-        <o-table-column label="Created"
-                        v-slot="{ row }">
-          <PrettyTime :value="row.created" />
-        </o-table-column>
-        <o-table-column label="Updated"
-                        v-slot="{ row }">
-          <PrettyTime :value="row.updated" />
-        </o-table-column>
-        <o-table-column label="Size"
-                        v-slot="{ row }">
-          <PrettyBytes :value="row.size" />
-        </o-table-column>
-        <o-table-column label="Actions"
-                        v-slot="{ row }">
-          <div style="display: flex; gap: 0.5rem;">
-
-            <a href="#" @click.prevent="showQtubeVideoDocument(row)">
-              <o-icon icon="eye" />
-              <span>View</span>
-            </a>
-
-            <a href="#" @click.prevent="qtubeVideoInitEdit(row)">
-              <o-icon icon="edit" />
-              <span>Edit</span>
-            </a>
-
-            <a href="#" @click.prevent="downloadResource(row)">
-              <o-icon icon="download" />
-              <span>Download</span>
-            </a>
-
-            <a href="#" @click.prevent="viewInQtube(row)">
-              <o-icon icon="external-link-alt" />
-              <span>View in Q-Tube</span>
-            </a>
-
-          </div>
-
-        </o-table-column>
-      </o-table>
-
-      <o-modal v-model:active="qtubeVideoShowDocument">
-        <div class="card">
-
-          <div class="card-header">
-            <div class="card-header-title">Q-Tube Video</div>
-          </div>
-
-          <div class="card-content"
-               v-if="qtubeVideoDocument">
-
-            <div class="columns">
-
-              <div class="column">
-
-                <h3 class="is-size-3">
-                  <span>{{ qtubeVideoDocument.title }}</span>
-                </h3>
-
-                <div v-if="qtubeVideoDocument.htmlDescription"
-                     v-html="qtubeVideoDocument.htmlDescription">
-                </div>
-                <div v-else>
-                  {{ qtubeVideoDocument.fullDescription }}
-                </div>
-
-                <div class="block"></div>
-
-                <o-field grouped>
-
-                <o-field label="Video Type">
-                  <span>{{ qtubeVideoDocument.videoType }}</span>
-                </o-field>
-
-                <o-field label="File Size">
-                  <PrettyBytes :value="qtubeVideoFileSize" />
-                </o-field>
-
-                </o-field>
-
-                <o-field label="Filename">
-                  <span>{{ qtubeVideoDocument.filename }}</span>
-                </o-field>
-
-                <o-field grouped>
-                  <o-field label="Created">
-                    <PrettyTime :value="qtubeVideoDocumentResource.created" />
-                  </o-field>
-                  <o-field label="Updated">
-                    <PrettyTime :value="qtubeVideoDocumentResource.updated" />
-                  </o-field>
-                </o-field>
-
-                <o-field label="Identifier">
-                  <PrettyIdentifier :value="qtubeVideoDocumentResource.identifier" />
-                </o-field>
-
-              </div>
-              <div class="column">
-
-                <QtubeVideoThumbnail v-if="qtubeVideoDocument.videoImage"
-                                     :video-image="qtubeVideoDocument.videoImage"
-                                     :frame-images="qtubeVideoDocument.extracts || []" />
-
-              </div>
-            </div>
-
-            <div class="block"
-                 style="display: flex; justify-content: space-between;">
-              <o-button @click="qtubeVideoShowJSON = !qtubeVideoShowJSON">
-                {{ qtubeVideoShowJSON ? "Hide" : "Show" }} JSON
-              </o-button>
-              <o-button variant="primary"
-                        icon-left="edit"
-                        @click="qtubeVideoInitEdit(qtubeVideoDocumentResource)">
-                Edit This
-              </o-button>
-            </div>
-
-            <div v-if="qtubeVideoShowJSON">
-              <VCodeBlock :code="JSON.stringify(qtubeVideoDocument, null, 2)" highlightjs />
-            </div>
-
-          </div>
+      <o-field>
+        <div style="display: flex; align-items: center; gap: 2rem;">
+          <o-button @click="qtubeVideosRefresh()"
+                    variant="primary"
+                    :disabled="qtubeVideosLoading || !showingName">
+            {{ qtubeVideosLoading ? "Working, please wait..." : (qtubeVideosData ? "Refresh" : "Load") }}
+          </o-button>
+          <span v-if="qtubeVideosData">
+            found {{ qtubeVideosDataFinal?.length }} result{{ qtubeVideosDataFinal?.length == 1 ? '' : 's' }}
+          </span>
         </div>
-      </o-modal>
+      </o-field>
 
-      <o-modal v-model:active="qtubeVideoEditing" cancelable="escape,x">
+      <QtubeVideoList :initial-data="qtubeVideosData"
+                      :loading="qtubeVideosLoading"
+                      @final-data="data => qtubeVideosDataFinal = data"
+                      @view-metadata="showQtubeVideoDocument" />
+
+      <QtubeVideoDialog ref="qtubeViewDialog"
+                        v-model:active="qtubeVideoShowDocument"
+                        @edit-metadata="info => qtubeVideoInitEdit(info, null, true)" />
+
+      <o-modal v-model:active="qtubeVideoEditing"
+               :cancelable="['escape','x']"
+               @close="qtubeVideoCancelEdit()">
         <div ref="qtubeEditTop"></div>
         <div class="card">
 
@@ -604,7 +586,7 @@ export default {
                           :disabled="qtubeVideoPublishing">
                   {{ qtubeVideoPublishing ? "Working, please wait..." : "Publish" }}
                 </o-button>
-                <o-button @click="qtubeVideoEditing = false">
+                <o-button @click="qtubeVideoCancelEdit()">
                   Cancel
                 </o-button>
               </div>
@@ -623,17 +605,52 @@ export default {
         <o-field label="Limit">
           <o-input v-model="videoResourcesLimit" type="number" />
         </o-field>
+        <o-field label="Q-Tube Filter">
+          <o-field grouped>
+            <o-select v-model="videoResourcesQtubeFilter"
+                      :disabled="!qtubeVideoCache">
+              <option :value="null">(no filter)</option>
+              <option value="qtube-only">show videos ALREADY in Q-Tube</option>
+              <option value="qtube-missing">show videos NOT YET in Q-Tube</option>
+            </o-select>
+            <o-button @click="rebuildQtubeCache()"
+                      :disabled="!showingName || qtubeVideoCacheBuilding">
+              {{ qtubeVideoCacheBuilding ? "Working, please wait..." : `${qtubeVideoCache ? 're-' : '' }build Q-Tube cache` }}
+            </o-button>
+          </o-field>
+        </o-field>
       </o-field>
 
       <div style="display: flex; align-items: center; gap: 2rem;">
         <o-button @click="videoResourcesRefresh()"
-                  :disabled="videoResourcesLoading">
+                  variant="primary"
+                  :disabled="videoResourcesLoading || !showingName">
           {{ videoResourcesLoading ? "Working, please wait..." : (videoResourcesData ? "Refresh" : "Load") }}
         </o-button>
         <span v-if="videoResourcesData">
           found {{ videoResourcesData.length }} result{{ videoResourcesData.length == 1 ? '' : 's' }}
         </span>
       </div>
+
+      <o-loading v-model:active="qtubeVideoCacheBuilding"
+                 full-page>
+        <div class="card" style="min-width: 50rem;">
+          <div class="card-content">
+            <p class="block">
+              Rebuilding cache, please wait...
+            </p>
+            <o-field label="Completed" horizontal>
+              <span>{{ qtubeVideoCacheBuildingStep }}</span>
+            </o-field>
+            <o-field label="Total" horizontal>
+              <span>{{ qtubeVideoCacheBuildingTotal }}</span>
+            </o-field>
+            <progress class="progress is-large"
+                      :value="qtubeVideoCacheBuildingStep"
+                      :max="qtubeVideoCacheBuildingTotal" />
+          </div>
+        </div>
+      </o-loading>
 
       <o-table v-if="videoResourcesData"
                :data="videoResourcesData || []"
@@ -650,6 +667,10 @@ export default {
         <o-table-column label="Identifier"
                         v-slot="{ row }">
           <PrettyIdentifier :value="row.identifier" />
+        </o-table-column>
+        <o-table-column label="Title"
+                        v-slot="{ row }">
+          {{ row.metadata?.title }}
         </o-table-column>
         <o-table-column label="Created"
                         v-slot="{ row }">
@@ -672,12 +693,13 @@ export default {
               <span>Metadata</span>
             </a>
 
-            <a href="#" @click.prevent="downloadResource(row)">
+            <a href="#" @click.prevent="$refs.downloader.downloadResource(row)">
               <o-icon icon="download" />
               <span>Download</span>
             </a>
 
-            <a href="#" @click.prevent="qtubeVideoInitEdit(null, row)">
+            <a v-if="qordialAuthStore.username && showingName == qordialAuthStore.username"
+               href="#" @click.prevent="qtubeVideoInitEdit(null, row)">
               <o-icon icon="upload" />
               <span>Publish to Q-Tube</span>
             </a>